Refactor cart service to improve code reusability
Related to #7

Refactor the shopping cart module to improve code reusability.

* **Add CartService class**
  - Create a new `CartService` class in `mall-portal/src/main/java/com/macro/mall/portal/service/CartService.java`
  - Add methods for common cart operations such as adding, updating, deleting, and clearing cart items

* **Update OmsCartItemServiceImpl class**
  - Remove common cart operations and delegate to `CartService`
  - Update methods to call `CartService` for common operations

* **Update OmsPortalOrderServiceImpl class**
  - Remove common cart operations and delegate to `CartService`
  - Update methods to call `CartService` for common operations

* **Update OmsCartItemController class**
  - Update imports to include `CartService`
  - Update methods to call `CartService` for common operations

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/sxthunder/mall/issues/7?shareId=XXXX-XXXX-XXXX-XXXX).
