+++ mall-portal/src/main/java/com/macro/mall/portal/controller/OmsCartItemController.java
@@ -4,7 +4,7 @@ import com.macro.mall.common.api.CommonResult;
 import com.macro.mall.model.OmsCartItem;
 import com.macro.mall.portal.domain.CartProduct;
 import com.macro.mall.portal.domain.CartPromotionItem;
-import com.macro.mall.portal.service.OmsCartItemService;
+import com.macro.mall.portal.service.CartService;
 import com.macro.mall.portal.service.UmsMemberService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -25,7 +25,7 @@ import java.util.List;
 @RequestMapping("/cart")
 public class OmsCartItemController {
     @Autowired
-    private OmsCartItemService cartItemService;
+    private CartService cartService;
     @Autowired
     private UmsMemberService memberService;
 
@@ -33,7 +33,7 @@ public class OmsCartItemController {
     @RequestMapping(value = "/add", method = RequestMethod.POST)
     @ResponseBody
     public CommonResult add(@RequestBody OmsCartItem cartItem) {
-        int count = cartItemService.add(cartItem);
+        int count = cartService.addCartItem(cartItem);
         if (count > 0) {
             return CommonResult.success(count);
         }
@@ -44,7 +44,7 @@ public class OmsCartItemController {
     @RequestMapping(value = "/list", method = RequestMethod.GET)
     @ResponseBody
     public CommonResult<List<OmsCartItem>> list() {
-        List<OmsCartItem> cartItemList = cartItemService.list(memberService.getCurrentMember().getId());
+        List<OmsCartItem> cartItemList = cartService.listCartItems(memberService.getCurrentMember().getId());
         return CommonResult.success(cartItemList);
     }
 
@@ -52,7 +52,7 @@ public class OmsCartItemController {
     @RequestMapping(value = "/list/promotion", method = RequestMethod.GET)
     @ResponseBody
     public CommonResult<List<CartPromotionItem>> listPromotion(@RequestParam(required = false) List<Long> cartIds) {
-        List<CartPromotionItem> cartPromotionItemList = cartItemService.listPromotion(memberService.getCurrentMember().getId(), cartIds);
+        List<CartPromotionItem> cartPromotionItemList = cartService.listCartItemsWithPromotion(memberService.getCurrentMember().getId(), cartIds);
         return CommonResult.success(cartPromotionItemList);
     }
 
@@ -61,7 +61,7 @@ public class OmsCartItemController {
     @ResponseBody
     public CommonResult updateQuantity(@RequestParam Long id,
                                        @RequestParam Integer quantity) {
-        int count = cartItemService.updateQuantity(id, memberService.getCurrentMember().getId(), quantity);
+        int count = cartService.updateCartItemQuantity(id, memberService.getCurrentMember().getId(), quantity);
         if (count > 0) {
             return CommonResult.success(count);
         }
@@ -72,7 +72,7 @@ public class OmsCartItemController {
     @RequestMapping(value = "/getProduct/{productId}", method = RequestMethod.GET)
     @ResponseBody
     public CommonResult<CartProduct> getCartProduct(@PathVariable Long productId) {
-        CartProduct cartProduct = cartItemService.getCartProduct(productId);
+        CartProduct cartProduct = cartService.getCartProduct(productId);
         return CommonResult.success(cartProduct);
     }
 
@@ -80,7 +80,7 @@ public class OmsCartItemController {
     @RequestMapping(value = "/update/attr", method = RequestMethod.POST)
     @ResponseBody
     public CommonResult updateAttr(@RequestBody OmsCartItem cartItem) {
-        int count = cartItemService.updateAttr(cartItem);
+        int count = cartService.updateCartItemAttr(cartItem);
         if (count > 0) {
             return CommonResult.success(count);
         }
@@ -91,7 +91,7 @@ public class OmsCartItemController {
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
     @ResponseBody
     public CommonResult delete(@RequestParam("ids") List<Long> ids) {
-        int count = cartItemService.delete(memberService.getCurrentMember().getId(), ids);
+        int count = cartService.deleteCartItems(memberService.getCurrentMember().getId(), ids);
         if (count > 0) {
             return CommonResult.success(count);
         }
@@ -102,7 +102,7 @@ public class OmsCartItemController {
     @RequestMapping(value = "/clear", method = RequestMethod.POST)
     @ResponseBody
     public CommonResult clear() {
-        int count = cartItemService.clear(memberService.getCurrentMember().getId());
+        int count = cartService.clearCart(memberService.getCurrentMember().getId());
         if (count > 0) {
             return CommonResult.success(count);
         }
