+++ mall-portal/src/main/java/com/macro/mall/portal/service/CartService.java
@@ -1,0 +1,107 @@
+package com.macro.mall.portal.service;
+
+import com.macro.mall.mapper.OmsCartItemMapper;
+import com.macro.mall.model.OmsCartItem;
+import com.macro.mall.model.OmsCartItemExample;
+import com.macro.mall.model.UmsMember;
+import com.macro.mall.portal.dao.PortalProductDao;
+import com.macro.mall.portal.domain.CartProduct;
+import com.macro.mall.portal.service.UmsMemberService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
+
+import java.util.Date;
+import java.util.List;
+
+@Service
+public class CartService {
+
+    @Autowired
+    private OmsCartItemMapper cartItemMapper;
+
+    @Autowired
+    private PortalProductDao productDao;
+
+    @Autowired
+    private UmsMemberService memberService;
+
+    public int addCartItem(OmsCartItem cartItem) {
+        int count;
+        UmsMember currentMember = memberService.getCurrentMember();
+        cartItem.setMemberId(currentMember.getId());
+        cartItem.setMemberNickname(currentMember.getNickname());
+        cartItem.setDeleteStatus(0);
+        OmsCartItem existCartItem = getCartItem(cartItem);
+        if (existCartItem == null) {
+            cartItem.setCreateDate(new Date());
+            count = cartItemMapper.insert(cartItem);
+        } else {
+            cartItem.setModifyDate(new Date());
+            existCartItem.setQuantity(existCartItem.getQuantity() + cartItem.getQuantity());
+            count = cartItemMapper.updateByPrimaryKey(existCartItem);
+        }
+        return count;
+    }
+
+    public List<OmsCartItem> listCartItems(Long memberId) {
+        OmsCartItemExample example = new OmsCartItemExample();
+        example.createCriteria().andDeleteStatusEqualTo(0).andMemberIdEqualTo(memberId);
+        return cartItemMapper.selectByExample(example);
+    }
+
+    public int updateCartItemQuantity(Long id, Long memberId, Integer quantity) {
+        OmsCartItem cartItem = new OmsCartItem();
+        cartItem.setQuantity(quantity);
+        OmsCartItemExample example = new OmsCartItemExample();
+        example.createCriteria().andDeleteStatusEqualTo(0)
+                .andIdEqualTo(id).andMemberIdEqualTo(memberId);
+        return cartItemMapper.updateByExampleSelective(cartItem, example);
+    }
+
+    public int deleteCartItems(Long memberId, List<Long> ids) {
+        OmsCartItem record = new OmsCartItem();
+        record.setDeleteStatus(1);
+        OmsCartItemExample example = new OmsCartItemExample();
+        example.createCriteria().andIdIn(ids).andMemberIdEqualTo(memberId);
+        return cartItemMapper.updateByExampleSelective(record, example);
+    }
+
+    public CartProduct getCartProduct(Long productId) {
+        return productDao.getCartProduct(productId);
+    }
+
+    public int updateCartItemAttr(OmsCartItem cartItem) {
+        //删除原购物车信息
+        OmsCartItem updateCart = new OmsCartItem();
+        updateCart.setId(cartItem.getId());
+        updateCart.setModifyDate(new Date());
+        updateCart.setDeleteStatus(1);
+        cartItemMapper.updateByPrimaryKeySelective(updateCart);
+        cartItem.setId(null);
+        addCartItem(cartItem);
+        return 1;
+    }
+
+    public int clearCart(Long memberId) {
+        OmsCartItem record = new OmsCartItem();
+        record.setDeleteStatus(1);
+        OmsCartItemExample example = new OmsCartItemExample();
+        example.createCriteria().andMemberIdEqualTo(memberId);
+        return cartItemMapper.updateByExampleSelective(record, example);
+    }
+
+    private OmsCartItem getCartItem(OmsCartItem cartItem) {
+        OmsCartItemExample example = new OmsCartItemExample();
+        OmsCartItemExample.Criteria criteria = example.createCriteria().andMemberIdEqualTo(cartItem.getMemberId())
+                .andProductIdEqualTo(cartItem.getProductId()).andDeleteStatusEqualTo(0);
+        if (cartItem.getProductSkuId() != null) {
+            criteria.andProductSkuIdEqualTo(cartItem.getProductSkuId());
+        }
+        List<OmsCartItem> cartItemList = cartItemMapper.selectByExample(example);
+        if (!CollectionUtils.isEmpty(cartItemList)) {
+            return cartItemList.get(0);
+        }
+        return null;
+    }
+}
